bug fixes
+ wrong type for pressio_thread_safety in io, metrics, and launch
+ various bug fixes for cusz
+ various fixes for smoke tests

+++ src/plugins/io/copy_template.cc
@@ -37,7 +37,7 @@ struct copy_template_io : public libpressio_io_plugin {
     pressio_options opts;
     opts.copy_from(impl->get_configuration());
     set(opts, "pressio:stability", "stable");
-    set(opts, "pressio:thread_safe",  static_cast<int32_t>(pressio_thread_safety_single));
+    set(opts, "pressio:thread_safe",  pressio_thread_safety_single);
     return opts;
   }
   virtual struct pressio_options get_documentation_impl() const override{
