+++ src/plugins/io/csv.cc
@@ -119,7 +119,7 @@ struct csv_io : public libpressio_io_plugin
   virtual struct pressio_options get_configuration_impl() const override{
     pressio_options opts;
     set(opts, "pressio:stability", "stable");
-    set(opts,"pressio:thread_safe",  static_cast<int32_t>(pressio_thread_safety_multiple));
+    set(opts,"pressio:thread_safe",  pressio_thread_safety_multiple);
     return opts;
   }
